Adding generation of redoc docs for APIs
