+++ apigenerator/resources/1 - Project/1 - BaseProject/Project/app.py
@@ -3,6 +3,7 @@ from src.e_Infra.g_Environment.EnvironmentVariables import *
 
 # Controller Imports #
 from src.a_Presentation.d_Swagger.SwaggerController import *
+from src.a_Presentation.c_Redoc.RedocController import *
 from src.a_Presentation.b_Custom.OptionsController import *
 from src.a_Presentation.b_Custom.SQLController import *
 from src.a_Presentation.b_Custom.BeforeRequestController import *
