+++ apigenerator/resources/1 - Project/1 - BaseProject/Project/config/redoc.html
@@ -15,7 +15,7 @@
   </head>
   <body>
     <!-- This will display the YAML content -->
-    <redoc spec-url='/redoc'></redoc>
+    <redoc spec-url='/redoc/spec'></redoc>
     <script src="https://cdn.redoc.ly/redoc/latest/bundles/redoc.standalone.js"></script>
   </body>
 </html>
