+++ apigenerator/resources/1 - Project/1 - BaseProject/Project/src/a_Presentation/c_Redoc/RedocController.py
@@ -1,10 +1,16 @@
-from flask import render_template
+from src.e_Infra.b_Builders.FlaskBuilder import *
 
 
-def build_redoc_blueprint(app_handler, redoc_blueprint):
-    @redoc_blueprint.route('/redoc')
-    def redoc():
-        return render_template('redoc.html')
+# Methods to create main Redoc docs #
+@redoc_blueprint.route('/redoc')
+def redoc():
+    return render_template('redoc.html')
 
-    # Register the blueprint
-    app_handler.register_blueprint(redoc_blueprint)
+
+@redoc_blueprint.route('/redoc/spec')
+def spec():
+    yaml_path = os.path.join(os.getcwd(), 'config', 'swagger.yaml')
+    with open(yaml_path, 'r') as yaml_file:
+        yaml_content = yaml_file.read()
+    return yaml_content
+
