+++ apigenerator/resources/1 - Project/1 - BaseProject/Project/src/e_Infra/b_Builders/FlaskBuilder.py
@@ -2,22 +2,18 @@
 import os
 
 # Importing Flask #
-from flask import Flask, Blueprint
+from flask import Flask, Blueprint, render_template
 
 # Infra Imports #
 from src.e_Infra.b_Builders.a_Swagger.SwaggerBuilder import *
-from src.a_Presentation.c_Redoc.RedocController import *
 
 
 # Initializing Flask #
-app_handler = Flask(__name__, template_folder=os.path.join(os.getcwd(), 'config'))
+app_handler = Flask(__name__)
+app_handler.template_folder = os.path.join(os.getcwd(), 'config')
 
 # Creating Redoc blueprint #
 redoc_blueprint = Blueprint('redoc', __name__)
 
 # Building Swagger Blueprint #
 build_swagger_blueprint(app_handler)
-
-
-# Building Redoc Blueprint #
-build_redoc_blueprint(app_handler, redoc_blueprint)
