+++ apigenerator/resources/5 - Redoc/templates/domain.html
@@ -15,7 +15,7 @@
   </head>
   <body>
     <!-- This will display the YAML content -->
-    <redoc spec-url='/meta_string'></redoc>
+    <redoc spec-url='/redoc/spec/meta_string'></redoc>
     <script src="https://cdn.redoc.ly/redoc/latest/bundles/redoc.standalone.js"></script>
   </body>
 </html>
